+++ gsw_xarray/_util.py
@@ -16,6 +16,7 @@ def submodule_attr_compat(submodule_name):
 
 def submodule_dir_compat(submodule_name):
     submodule = import_module(submodule_name)
-
-    def dir_(name):
+    print(submodule, submodule_name)
+    def dir_():
         return dir(submodule)
+    return dir_
